So much stuff.T Thoughts on the Nearest Neighbor problem
The sort of the distances really needs to be sorting the indices of the distances using the distance of the index.  The SortAndTake I helped you with
does not do that but it could be modified to do it.  Basically
Create and new version of SortAndTake to SortIndexAndTake
Change the (when(io.newInputs) to unitialize the sortReg with indices instead of input values
Declare sortReg to be UInts (best for indices) instead of Fixed
val sortReg      = Reg(Vec(inputSize, UInt(32.W)))
Initialize them when new inputs come in to be just the indices
sortReg.zipWithIndex.foreach { case (reg, index) => reg := index.U }
In the reg comparisons in SortAndTake compare
Use when(io.inputs(regA) < io.inputs(regB)) instead of when(regA > regB)
this way the outputs vector will contain the indices of the lowest value functions
I don't think you should put your distance calculations and parsing and everything inside the SortAndTake.
Inside a Hardware nearest neighbor you need to create an instance of SortIndexAndTake,
and load it's values and monitor sortDone to know when the sort if finished

When it's done you can then find the most frequently occurring
(I think that's topKClasses.groupBy(identity).mapValues(_.size).maxBy(_._2)._1 does)
That's a little tricky but you should be able to do it
creating an Vec of counts and another Vec of that maps your results to the count Vec
I think you can probably use foldLeft somehow on the indices (using your Vec of count to choose the index with the highest count)

All this is going to require that your HardwareNearestNeighbours module will need some sort of enable and done flags, because
the sorts will take multiple cycles to complete.  (that's pretty standard with something this complicated).